OneDrive/Desktop/AutoSense/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
celery==5.3.1
redis==5.0.0